Remove leaking of OS abstractions
#2901 was good, but we can do even better.
